requirements.txt
flask
openai
nltk
python-dotenv
MarkupSafe
jinja2